Add main assembler file.
The Makefile will generate assembler files from the PBM images, concatenate them with the main source file, and assemble. We don't have a linker for the 6502 code, so this is the best we can do. The demo just shows a splash screen, then a graphic screen in green, then the splash screen again in yellow, then scrolls it away to the right. The final loop never terminates and requires a RESET to get back to the monitor.