Add the final checksum HEX file.
The HEX file is not normally included in the repo, but I've added it here so that other UK101 users can run the demo without having to assemble it. But there are two obstacles: it requires the Extended Monitor's checksum loader, and it overwrites the monitor in RAM. The demo is now big enough that it also requires more than 4k of RAM, but that's just a matter of plugging in more 2114 RAM chips. No, the real difficulty is the Extended Monitor, which in my machine is in EPROM. Two solutions: write a checksum loader that resides at the top of the 8k RAM area. Or, modify the assembly-language source of the demo to avoid overwriting the monitor.